refactor. RacingCars 일급 컬렉션을 생성해 관련된 행위를 관리하도록 변경
 - RacingCarGame에서는 로직의 흐름만 관리하도록 변경
